+++ packages/bluesky/lib/src/params/generator_param.freezed.dart
@@ -26,6 +26,7 @@ mixin _$GeneratorParam {
   List<Facet>? get descriptionFacets => throw _privateConstructorUsedError;
   Blob? get avatar => throw _privateConstructorUsedError;
   DateTime? get createdAt => throw _privateConstructorUsedError;
+  Map<String, dynamic> get unspecced => throw _privateConstructorUsedError;
 
   Map<String, dynamic> toJson() => throw _privateConstructorUsedError;
   @JsonKey(ignore: true)
@@ -45,7 +46,8 @@ abstract class $GeneratorParamCopyWith<$Res> {
       String? description,
       List<Facet>? descriptionFacets,
       Blob? avatar,
-      DateTime? createdAt});
+      DateTime? createdAt,
+      Map<String, dynamic> unspecced});
 
   $BlobCopyWith<$Res>? get avatar;
 }
@@ -69,6 +71,7 @@ class _$GeneratorParamCopyWithImpl<$Res, $Val extends GeneratorParam>
     Object? descriptionFacets = freezed,
     Object? avatar = freezed,
     Object? createdAt = freezed,
+    Object? unspecced = null,
   }) {
     return _then(_value.copyWith(
       did: null == did
@@ -95,6 +98,10 @@ class _$GeneratorParamCopyWithImpl<$Res, $Val extends GeneratorParam>
           ? _value.createdAt
           : createdAt // ignore: cast_nullable_to_non_nullable
               as DateTime?,
+      unspecced: null == unspecced
+          ? _value.unspecced
+          : unspecced // ignore: cast_nullable_to_non_nullable
+              as Map<String, dynamic>,
     ) as $Val);
   }
 
@@ -125,7 +132,8 @@ abstract class _$$_GeneratorParamCopyWith<$Res>
       String? description,
       List<Facet>? descriptionFacets,
       Blob? avatar,
-      DateTime? createdAt});
+      DateTime? createdAt,
+      Map<String, dynamic> unspecced});
 
   @override
   $BlobCopyWith<$Res>? get avatar;
@@ -148,6 +156,7 @@ class __$$_GeneratorParamCopyWithImpl<$Res>
     Object? descriptionFacets = freezed,
     Object? avatar = freezed,
     Object? createdAt = freezed,
+    Object? unspecced = null,
   }) {
     return _then(_$_GeneratorParam(
       did: null == did
@@ -174,6 +183,10 @@ class __$$_GeneratorParamCopyWithImpl<$Res>
           ? _value.createdAt
           : createdAt // ignore: cast_nullable_to_non_nullable
               as DateTime?,
+      unspecced: null == unspecced
+          ? _value._unspecced
+          : unspecced // ignore: cast_nullable_to_non_nullable
+              as Map<String, dynamic>,
     ));
   }
 }
@@ -188,8 +201,10 @@ class _$_GeneratorParam implements _GeneratorParam {
       this.description,
       final List<Facet>? descriptionFacets,
       this.avatar,
-      this.createdAt})
-      : _descriptionFacets = descriptionFacets;
+      this.createdAt,
+      final Map<String, dynamic> unspecced = emptyJson})
+      : _descriptionFacets = descriptionFacets,
+        _unspecced = unspecced;
 
   factory _$_GeneratorParam.fromJson(Map<String, dynamic> json) =>
       _$$_GeneratorParamFromJson(json);
@@ -215,10 +230,18 @@ class _$_GeneratorParam implements _GeneratorParam {
   final Blob? avatar;
   @override
   final DateTime? createdAt;
+  final Map<String, dynamic> _unspecced;
+  @override
+  @JsonKey()
+  Map<String, dynamic> get unspecced {
+    if (_unspecced is EqualUnmodifiableMapView) return _unspecced;
+    // ignore: implicit_dynamic_type
+    return EqualUnmodifiableMapView(_unspecced);
+  }
 
   @override
   String toString() {
-    return 'GeneratorParam(did: $did, displayName: $displayName, description: $description, descriptionFacets: $descriptionFacets, avatar: $avatar, createdAt: $createdAt)';
+    return 'GeneratorParam(did: $did, displayName: $displayName, description: $description, descriptionFacets: $descriptionFacets, avatar: $avatar, createdAt: $createdAt, unspecced: $unspecced)';
   }
 
   @override
@@ -235,7 +258,9 @@ class _$_GeneratorParam implements _GeneratorParam {
                 .equals(other._descriptionFacets, _descriptionFacets) &&
             (identical(other.avatar, avatar) || other.avatar == avatar) &&
             (identical(other.createdAt, createdAt) ||
-                other.createdAt == createdAt));
+                other.createdAt == createdAt) &&
+            const DeepCollectionEquality()
+                .equals(other._unspecced, _unspecced));
   }
 
   @JsonKey(ignore: true)
@@ -247,7 +272,8 @@ class _$_GeneratorParam implements _GeneratorParam {
       description,
       const DeepCollectionEquality().hash(_descriptionFacets),
       avatar,
-      createdAt);
+      createdAt,
+      const DeepCollectionEquality().hash(_unspecced));
 
   @JsonKey(ignore: true)
   @override
@@ -270,7 +296,8 @@ abstract class _GeneratorParam implements GeneratorParam {
       final String? description,
       final List<Facet>? descriptionFacets,
       final Blob? avatar,
-      final DateTime? createdAt}) = _$_GeneratorParam;
+      final DateTime? createdAt,
+      final Map<String, dynamic> unspecced}) = _$_GeneratorParam;
 
   factory _GeneratorParam.fromJson(Map<String, dynamic> json) =
       _$_GeneratorParam.fromJson;
@@ -287,6 +314,8 @@ abstract class _GeneratorParam implements GeneratorParam {
   Blob? get avatar;
   @override
   DateTime? get createdAt;
+  @override
+  Map<String, dynamic> get unspecced;
   @override
   @JsonKey(ignore: true)
   _$$_GeneratorParamCopyWith<_$_GeneratorParam> get copyWith =>