+++ packages/bluesky/lib/src/params/generator_param.g.dart
@@ -29,6 +29,13 @@ _$_GeneratorParam _$$_GeneratorParamFromJson(Map json) => $checkedCreate(
                   : Blob.fromJson(Map<String, Object?>.from(v as Map))),
           createdAt: $checkedConvert('createdAt',
               (v) => v == null ? null : DateTime.parse(v as String)),
+          unspecced: $checkedConvert(
+              'unspecced',
+              (v) =>
+                  (v as Map?)?.map(
+                    (k, e) => MapEntry(k as String, e),
+                  ) ??
+                  emptyJson),
         );
         return val;
       },
@@ -51,5 +58,6 @@ Map<String, dynamic> _$$_GeneratorParamToJson(_$_GeneratorParam instance) {
       instance.descriptionFacets?.map((e) => e.toJson()).toList());
   writeNotNull('avatar', instance.avatar?.toJson());
   writeNotNull('createdAt', instance.createdAt?.toIso8601String());
+  val['unspecced'] = instance.unspecced;
   return val;
 }