+++ packages/bluesky/lib/src/params/strong_ref_param.g.dart
@@ -18,6 +18,13 @@ _$_StrongRefParam _$$_StrongRefParamFromJson(Map json) => $checkedCreate(
               'uri', (v) => atUriConverter.fromJson(v as String)),
           createdAt: $checkedConvert('createdAt',
               (v) => v == null ? null : DateTime.parse(v as String)),
+          unspecced: $checkedConvert(
+              'unspecced',
+              (v) =>
+                  (v as Map?)?.map(
+                    (k, e) => MapEntry(k as String, e),
+                  ) ??
+                  emptyJson),
         );
         return val;
       },
@@ -36,5 +43,6 @@ Map<String, dynamic> _$$_StrongRefParamToJson(_$_StrongRefParam instance) {
   }
 
   writeNotNull('createdAt', instance.createdAt?.toIso8601String());
+  val['unspecced'] = instance.unspecced;
   return val;
 }