+++ packages/bluesky/lib/src/params/thread_param.g.dart
@@ -28,6 +28,13 @@ _$_ThreadParam _$$_ThreadParamFromJson(Map json) => $checkedCreate(
               (v) => (v as List<dynamic>?)?.map((e) => e as String).toList()),
           createdAt: $checkedConvert('createdAt',
               (v) => v == null ? null : DateTime.parse(v as String)),
+          unspecced: $checkedConvert(
+              'unspecced',
+              (v) =>
+                  (v as Map?)?.map(
+                    (k, e) => MapEntry(k as String, e),
+                  ) ??
+                  emptyJson),
         );
         return val;
       },
@@ -51,6 +58,7 @@ Map<String, dynamic> _$$_ThreadParamToJson(_$_ThreadParam instance) {
           instance.embed, embedConverter.toJson));
   writeNotNull('languageTags', instance.languageTags);
   writeNotNull('createdAt', instance.createdAt?.toIso8601String());
+  val['unspecced'] = instance.unspecced;
   return val;
 }